Fix tests and mistakes found from tests
