+++ src/code42cli/cmds/alert_rules.py
@@ -12,9 +12,9 @@ from code42cli.cmds.shared import get_user_id
 from code42cli.errors import Code42CLIError
 from code42cli.errors import InvalidRuleTypeError
 from code42cli.file_readers import read_csv_arg
+from code42cli.options import format_option
 from code42cli.options import OrderedGroup
 from code42cli.options import sdk_options
-from code42cli.options import format_option
 
 
 class AlertRuleTypes:
