+++ src/code42cli/cmds/legal_hold.py
@@ -15,9 +15,9 @@ from code42cli.errors import LegalHoldNotFoundOrPermissionDeniedError
 from code42cli.errors import UserAlreadyAddedError
 from code42cli.errors import UserNotInLegalHoldError
 from code42cli.file_readers import read_csv_arg
+from code42cli.options import format_option
 from code42cli.options import OrderedGroup
 from code42cli.options import sdk_options
-from code42cli.options import format_option
 from code42cli.util import format_string_list_to_columns
