+++ tests/test_output_formats.py
@@ -1,11 +1,9 @@
 import json
 from collections import OrderedDict
 
-from code42cli.output_formats import extraction_output_format
 from code42cli.output_formats import get_dynamic_header
 from code42cli.output_formats import output_format
 from code42cli.output_formats import to_csv
-from code42cli.output_formats import to_dynamic_csv
 from code42cli.output_formats import to_formatted_json
 from code42cli.output_formats import to_json
 from code42cli.output_formats import to_table
@@ -175,21 +173,6 @@ def test_output_format_returns_to_csv_function_when_csv_format_option_is_passed(
 def test_output_format_returns_to_table_function_when_no_format_option_is_passed():
     format_function = output_format(None, None, None)
     assert id(format_function) == id(to_table)
-
-
-def test_output_format_returns_to_dynamic_csv_function_when_csv_option_is_passed():
-    extraction_output_format_function = extraction_output_format(None, None, "CSV")
-    assert id(extraction_output_format_function) == id(to_dynamic_csv)
-
-
-def test_extraction_output_format_returns_to_formatted_json_function_when_json__option_is_passed():
-    format_function = extraction_output_format(None, None, "JSON")
-    assert id(format_function) == id(to_formatted_json)
-
-
-def test_extraction_output_format_returns_to_json_function_when_raw_json_format_option_is_passed():
-    format_function = extraction_output_format(None, None, "RAW-JSON")
-    assert id(format_function) == id(to_json)
 
 
 def test_get_format_header_returns_all_keys_only_which_are_not_nested():
